do not delete failed repeatable jobs

+++ src/Yiiq.php
@@ -290,9 +290,10 @@ class Yiiq extends \CApplicationComponent
         $this->getPools()->executing->remove($id);
 
         if (
-            empty($this->faultIntervals)
+            (empty($this->faultIntervals)
             ||($this->faultIntervals
-            && $metadata->faults > count($this->faultIntervals))
+            && $metadata->faults > count($this->faultIntervals)))
+            && $metadata->type != self::TYPE_REPEATABLE
         ) {
             $metadata->save(true);
             $this->delete($id, false);
